make tests compilation easy

+++ tests/test-hiscoa.c
@@ -1,10 +1,11 @@
-#include "std.h"
-#include "hiscoa-compress.h"
+#include "../src/std.h"
+#include "../src/hiscoa-compress.h"
 #include "hiscoa-decompress.h"
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/types.h>
 
 static const unsigned band_size = 70;
 
@@ -64,7 +65,7 @@ int main(int argc, char **argv)
 			size_t bl = s / (width / 8);
 			size_t bs = hiscoa_compress_band(bandbuf, 2 * bsize,
 					buf, width / 8, bl,
-					(s < bsize) ? HISCOA_EOB_LAST : HISCOA_EOB_NORMAL,
+					(s < (int)bsize) ? HISCOA_EOB_LAST : HISCOA_EOB_NORMAL,
 					&params);
 			memset(dcbuf, 0xAA, bsize);
 			tmps = bs;
@@ -81,7 +82,7 @@ int main(int argc, char **argv)
 				fprintf(stderr, "! band compressed INCORRECTLY\n");
 				bad = true;
 			}
-			if (dcsize != s) {
+			if ((int)dcsize != s) {
 				fprintf(stderr, "! band size incorrect: %u instead of %u\n",
 						(unsigned) dcsize, (unsigned) s);
 				bad = true;
@@ -89,7 +90,7 @@ int main(int argc, char **argv)
 			if (bad)
 				errors += 1;
 		}
-		if (s < bsize)
+		if (s < (int)bsize)
 			break;
 	}
